Rename selectProofsRGLI.ts file to SelectProofs.ts
